deployments/base_requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
colorlog==6.8.2
Pillow==10.1.0
pydantic==2.9.2
transformers==4.45.1